added conditional checks to get around some errors when doing cleanup

+++ container/clean-all-containers.sh
@@ -1,5 +1,20 @@
 #! /bin/bash
 
 echo "cleaning all containers..."
-docker rm -f $(docker ps -a -q)
+
+if [ ! -z $(docker ps -q) ]
+then 
+  docker stop $(docker ps -q)
+  echo "stoping all containers"
+else
+  echo "no containers to stop"
+fi
+
+if [ ! -z $(docker ps -a -q) ]
+then
+  docker rm -f $(docker ps -a -q)
+else
+  echo "no containers to clean"
+fi
+
 echo "finished cleaning containers"
